Pass test, don't implement Stateful, and don't set Prior Loan Payment
Since the LoanServicer was Stateful, it was saving the
loanPaymentsToCreate between batch jobs (the list grew by a few thousand
LPs each batch). Simple fix to remove the Stateful implementation, which
wasn't really needed in the first place. Also retired the Prior Loan
Payment field since it isn't being used. we can also add it back in when
needed.