Preview Bills Window
Lightning Component and Visualforce Page :
When the "Preview Bills" button is hit on a Production Update, this
finds all related UASes and
calculates what their bill will be (NMCs allocated and then with 10%
discount). It puts that list in a table.

Another button "Generate Bills" on the Production Update runs what used
to happen on the insertion of production update.

Process will change to: 1.) Create Production Update when we get a new
host account bill 2.) Preview and Review what the bills will be over a
longer period 3.) Generate Bills on the 15th

Secondly, I reused the component for a button on the account so the BWOC
can pull a list of all current and future bills for any given customer.
Also pulled a list of transactions, because they currently login to the
person's portal to see that.