Adjustments to the Tool Tip
Edited the tool tip component so it can be flipped to the left side of
its anchor. The tip was running of the modal's page so the user couldn't
read all of it as it was getting cut off. It defaults to the right side,
but in development you can set the placement attribute to left.