Edits to Attachment Upload
Update of Equipment object is not dependent on the user clicking the
"Done" button. Now it is fired directly after the file upload so if they
refresh the page it still updates the record. Instead of re-using the
SLPCustomerController to update the resi equipment I used the
updateSObject method on the BlueWave Parent, so the FileUploadWindow can
save any SObject not just opportunities or Resi Equipment records. (This
means I un-did the edits I made to the SLPCustomerHandler

I want to move the saveObject method on the SLPFileUploadWindowHelper to
just reference BlueWaveParent but I keep getting the same error. If I
can find a way to get around that I will just reference
helper.saveSObject instead of using that block.