Recurring Orders Stop for Cancelled Customers
Sets the Payment Status to stopped, charge amount =0 and leaves a note
when it tries to re-assign an Account Bill to the next months Account
Bill and that bill does not exist. This will prevent cancelled customers
(ie no more bills) from having recurring orders that re-charge the same
amount every month.