ACH Signup From DocuSign Hitting Error
Fields on the opportunity are populated from DocuSign when a CS customer
fills out the ACH fields on a contract. This is an old class that moves
autopay info from the Opportunity to a Chargent Order, and removes the
data from the Opportunity. This was hitting an error because the gateway
was pushing a SSS id to the Shipping last name field instead of an
entity ID.