Document Send Via Email
Instead of bills being generated and sent via email in one step, conga
now will just generate the document, we will review them and then we
will send out via email (apex).

Process:
1. Instead of just saving as an attachment, Conga now will save as a
Content (shows up in File list and Attachment list). Once the document
is generated, Conga checks the "Ready for Email Send" checkbox on the
Account Bill.
2. Link to the file will be generated on insertion of the documents by
apex (link is used in the portal and in the email)
3. Once the pdfs have been QCed, a user will hit a button on the Conga
Conductor record (the same thing that fires document generation) to send
the documents via email
4. Once the emails are sent, the "Published" checkbox on the Account
Bill is marked, which marks the Account Bill to be accessed by the
portal