Field on the Account needs to update as Total Due on the Account Bill changes
The directions to write checks field on the account is used for the
paper and email bill templates. We used to update this field every month
with dataloader before we sent out bills because even if it was
automatic, the values change a lot over the course of bill day. Its the
summary of total due, broken out by project company.

Total Due is a rollup field from the System Bills. Any time a system
bill is edited, it recalculates total due, which in turn triggers an
update of the Account Bill.  Total Due changes when a new production
update is logged (ie new System Bill is made), late fees are waived,
refunds or credits are applied. It does not change on payments.

>
https://developer.salesforce.com/docs/atlas.en-us.apexcode.meta/apexcode/apex_triggers_order_of_execution.htm