Updates to pass tests and removal of productsAssignmentHandler
Updates to pass tests and removal of productsAssignmentHandler

In order to remove the product assignment handler, the loan record warehouse had to be reconfigured. The two lead creation method shad to insert the products within them so that the leads could be assiged ot a product.  They were originally being assigned via the productassignmenthandler. The method also had to be removed from the testclsases that would call on the getProducts methods of the warehouse. Instead, they just call on the get leads methods which insert the products.