Disbursal update fixes (#226)
* Adjustments to better update disbursal records

The disbursal records were getting set to "Not yet disbursed" occasionally because Ops was handling the opp record differently than anticipated. The date and type were being updated in separate transactions, thus causing the disbursal to be set to not yet disbursed when the type was updated because the date field's oldValue and newValue weren't different from one another during that transaction. Updated the code to only set it to NYD when date is changed to a null value. Also removed unecessary spacing.

* Adjustments to DisbursalHandlerTestClass to comepnsate for last commit changes

Adjustments to DisbursalHandlerTestClass to comepnsate for last commit changes

* Build update