Move Cases from Lead > Contact & Account (#295)
* Move Cases from Lead > Contact & Account

Interested to hear other opinions here. Because this is just for record
keeping, I made it a scheduled job that we can set to run once, maybe
twice a day to catch any recently converted leads. So much happens at
lead conversion (and differs between CS/RL) I figured it was simpler
this way (don't want this to hold up lead conversion if for some reason
it fails) but I could also trigger it off the conversion itself.

* API Version 40

* Change to Be Triggered at Lead Conversion not Scheduled

* format if statement