Mar 15 Release (#307)
* Add ARL Test, amort schedule should pass, but fails

* CC Additional Email Field on CS Bills (#296)

* CC Additional Email Field on CS Bills

* Change Field to Lookup to Contact instead of second email

* _

* Set non-adjustable monthly rate! But get Internal Salesforce Error...

* ARL test passes, IBLS with paydown doesn't

* Internal Salesforce Error resolved by setting scale, LoanServicer tests pass!

* Percents need to be divided by 100

* Remove debugging

* Don't need Amort_PI_Due__c

* Add XIRR calc and test

* Add CC Functionality to Ch. Order Creation (#297)

Fields on Opp are populated by Docusign, need to be moved to Chargent
Order

* Set APR on Loan and Opp, doesn't handle product changes

* build.xml and missing .cls-meta.xml

* Always start the cashflow on 1/1 (apparently different result on 3/1 vs 2/28)

* Avoid SOQL limit issue

* Avoid SOQL limit issue

* Missing .cls-meta.xml

* Gross up DTI in Loan Underwriting Screen

* Gross up DTI in PCRApprovalHandler

* Gross up DTI in LoanUnderwriting snapshot if necessary

* Add adjustable fields to Loan Data table

* Refactor PCRApprovalHandler; need to add more DOER tests

* Don't include TCO and update Year 2/4 labels

* Break up testPCRApproval into two to avoid LimitException on emails

* Typo!

* Remove Total_Cost_of_Ownership

* Improve DOER PCR tests

* Test for low-fico handling, too.

* Switch from Product Code to Product Name

* Product Selection was using Product.Name, not ProductCode

* build.xml

* Default DOER floor to 681

* Make SOQL formatting consistent

* Return External Name in getCustomerProducts

* Return External Name in getCustomerProducts

* Co-Branding Logic (#304)

* Cobranding Logic

Updates 2 fields on the account, based on criteria from the most recent
account bill:
1.) Address to send checks - field on the entity that says if checks for
that project company are supposed to be sent to a lockbox, or address
other than 137 Newbury (this is currently dataloaded in)
2.) Client owner branding logo - name of the file that holds the logo
for the client - pulled into conga template

** both have similar logic. If a customer is cross-subscribed to an
entity that has something special, populate it. Unless it is on another
project that also has something special, in which case should go to
default value

(ie, if on Fairhaven and Barrett, check address should be 137 Newbury
not the Fairhaven lockbox. But if only on Fairhaven, it should use the
lockbox address. For logos, if on an AMP project alone, use the AMP
logo. If on AMP and Ameresco, use the default Bluewave logo.)

* _

* String

* ,

* test

* _

* String > ID

* Edits from Peter's comments

* build.xml

* Veteran Partner Task (#308)

Use Employed, Self_Employed, Retired, and Veteran checkboxes to process income docs

* Formatting cs portal (#305)

* Cobranding Logic

Updates 2 fields on the account, based on criteria from the most recent
account bill:
1.) Address to send checks - field on the entity that says if checks for
that project company are supposed to be sent to a lockbox, or address
other than 137 Newbury (this is currently dataloaded in)
2.) Client owner branding logo - name of the file that holds the logo
for the client - pulled into conga template

** both have similar logic. If a customer is cross-subscribed to an
entity that has something special, populate it. Unless it is on another
project that also has something special, in which case should go to
default value

(ie, if on Fairhaven and Barrett, check address should be 137 Newbury
not the Fairhaven lockbox. But if only on Fairhaven, it should use the
lockbox address. For logos, if on an AMP project alone, use the AMP
logo. If on AMP and Ameresco, use the default Bluewave logo.)

* _

* String

* ,

* test

* _

* Nav Bar and Dashboard Formatting

* Footer

* Formatting

* Updates

* Formatting

* _

* Revert "_"

This reverts commit e54c2362bff9b2c1a2d1a7d863b092bf274cc8fd.

* dynamic logo

* Testclass

* Static Resources

* Revert - move into folder

* UI Edit From Jenni

Remove production and solar farm values

* delete

* Revert Deletion of NavHeaderController Used by SLPHeader

* build

* tokens

* client brand key

* don't override build

* _

* Add CommunitySolarCustomerPortalTest to test runs

* build.xml

* build.xml