DOB Validation
Mock out of the box lightning error handling

Other input fields on this page are using out the box

<Lightning:Input> error handling, which hightlights the input field in a red outline and puts a message below the field.

We don't want to use lightning input for date, because the only option is a calendar click menu, which we've heard people don't like because you can't tab through it. So I use the BlueWaveDate component we use elsewhere but make the error message match the others (below the field)

We could also modify all the inputs to go off our custom components. Not sure what is better here.