Cannot Change Total Due on Published Bills
All the fields that should not be modified  should not recalculate every time the SB is updated:

Total Due = Due This Month + Late Fees + Adjustments + Overpayments

Need to figure out a way to handle Overpayments since that is a rollup field. I think it should be a validation rule?

None of those fields should be editable after bill is published

Only fields that should be editable are the balance fields, which can change as payments come in (Balance Net Late Payments, Carry Over Balance etc)