W-024452: Calculate Production Details for Debit Memos before Credit Memos (#4052)
For instance, this CM (a925b0000001dR6AAI) existed before the DM (a925b0000001lpHAAQ).

When P2B runs, it knows that it needs to generate PDs for the CM since "Ready for Production Detail" is checked until the CM is fully applied. The P2BService tries to, sees that the CM is now applied in Zuora (to the DM) and attempts to create PDs for it.

Because the DM doesn't have PDs yet, that results in no records (but no errors either), and because the CM is now fully applied in Zuora, P2BService unchecks "Ready for Production Detail" and no longer attempts to generate PDs for the CM.

This still leaves room for error if a DM and a Payment that it references via Returned Payment Part Id are created on the same day. That's unlikely since payments aren't usually returned on the same day. The reconciliation job planned in W-024453 should catch that case.