W-024454: Increase Production Detail Amount precision in apex (#4068)
For example, this Transfer (a900a000000GnhEAAS) has 13 Transfer PDs. This DM JE (a920a0000007enVAAQ) is related to it, so we distribute the $.09 between the 13 parts. Because PDs only have two decimals, it looks like it's missing $.02. So subsequent transactions against it, like CMs or Payments, miss $.02 as well.